+++ src/components/index.js
@@ -1,9 +1,9 @@
 // Commons Components
-export { default as Confirmation } from './Confirmation';
-export { default as Footer } from './Footer';
-export { default as Navbar } from './Navbar';
-export { default as Skeleton } from './Skeleteon';
-export { default as SectionTitle } from './SectionTitle';
+export { default as Confirmation } from './commons/Confirmation';
+export { default as Footer } from './commons/Footer';
+export { default as Navbar } from './commons/Navbar';
+export { default as Skeleton } from './commons/Skeleteon';
+export { default as SectionTitle } from './commons/SectionTitle';
 export { default as Breadcrumb } from './commons/Breadcrumb';
 // Auth Components
 export { default as SignUpForm } from './auth/SignUpForm';
